decoupled game from engine, redo turntaking a bit
redone turntaking also means that clients can disconnect midgame now.
decoupled game from engine means we can keep an entire game instance
server side and not just a world, which we can send netmessages to as if
it was a client, and that way keep things synced up a lot easier.
probably still some cleaning left, but still works so i mean that's
something.

oh, and added planet harvesting again